fixed:工序质量资料导入接口

+++ mes-common/common-model/src/main/java/com/richfit/mes/common/model/base/Sequence.java
@@ -42,12 +42,12 @@ public class Sequence extends BaseEntity<Sequence> {
     /**
      * 准结工时
      */
-    @ApiModelProperty(value = "准结工时", dataType = "int")
+    @ApiModelProperty(value = "准结工时", dataType = "float")
     private float prepareEndHours;
     /**
      * 额定工时
      */
-    @ApiModelProperty(value = "额定工时", dataType = "int")
+    @ApiModelProperty(value = "额定工时", dataType = "float")
     private float singlePieceHours;
     /**
      * 技术要求
